rework entry files for libs and bins
