+++ rust/shell/src/lib.rs
@@ -1,0 +1,10 @@
+#[cfg(any(feature = "appcsr", feature = "appssr"))]
+pub mod app;
+
+#[cfg(feature = "appcsr")]
+#[wasm_bindgen::prelude::wasm_bindgen]
+pub fn hydrate() {
+    use crate::app::*;
+    console_error_panic_hook::set_once();
+    leptos::mount::hydrate_body(App);
+}
