+++ rust/shell/src/main-app.rs
@@ -1,0 +1,47 @@
+mod error;
+
+#[cfg(feature = "db")]
+mod db;
+
+#[cfg(feature = "redis")]
+mod redis;
+
+#[cfg(feature = "rmqpub")]
+mod rmqpub;
+
+#[cfg(any(feature = "appssr", feature="appcsr"))]
+mod app;
+
+// FULLSTACK WEBAPP using leptos
+#[cfg(feature = "appssr")]
+#[tokio::main]
+async fn main() {
+    use axum::Router;
+    use leptos::logging::log;
+    use leptos::prelude::*;
+    use leptos_axum::{generate_route_list, LeptosRoutes};
+    use shell::app::*;
+
+    let conf = get_configuration(None).unwrap();
+    let addr = conf.leptos_options.site_addr;
+    let leptos_options = conf.leptos_options;
+    // Generate the list of routes in your Leptos App
+    let routes = generate_route_list(App);
+
+    let app = Router::new()
+        .leptos_routes(&leptos_options, routes, {
+            let leptos_options = leptos_options.clone();
+            move || shell(leptos_options.clone())
+        })
+        .fallback(leptos_axum::file_and_error_handler(shell))
+        .with_state(leptos_options);
+
+    // run our app with hyper
+    // `axum::Server` is a re-export of `hyper::Server`
+    log!("listening on http://{}", &addr);
+    let listener = tokio::net::TcpListener::bind(&addr).await.unwrap();
+    axum::serve(listener, app.into_make_service())
+        .await
+        .unwrap();
+}
+
