+++ rust/shell/src/main-worker.rs
@@ -1,11 +1,19 @@
+mod error;
+
+#[macro_use]
+extern crate rocket;
+
 #[cfg(feature = "db")]
 mod db;
-mod error;
+
 #[cfg(feature = "redis")]
 mod redis;
+
 #[cfg(feature = "rmqsub")]
 mod rmqsub;
 
+
+// RMQ subscriber worker using lapin
 #[tokio::main]
 #[cfg(feature = "rmqsub")]
 async fn main() {
@@ -22,5 +30,3 @@ async fn main() {
         .expect("couldnt start rmq's consumer");
 }
 
-#[cfg(not(feature = "rmqsub"))]
-fn main() {}
